% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsrch.R
\name{more_thuente}
\alias{more_thuente}
\title{More'-Thuente Line Search}
\usage{
more_thuente(c1 = 1e-04, c2 = 0.1, max_fn = Inf)
}
\arguments{
\item{c1}{Constant used in sufficient decrease condition. Should take a value
between 0 and 1.}

\item{c2}{Constant used in curvature condition. Should take a value between
\code{c1} and 1.}

\item{max_fn}{Maximum number of function evaluations allowed.}
}
\value{
Line search function for use in the conjugate gradient routine.
}
\description{
Line Search Factory Function
}
\details{
Returns a line search function that can be used in the
 \code{\link{conj_grad}} function. This uses a variant of the More-Thuente
 line search originally implemented in
 \href{http://www.netlib.org/minpack/}{MINPACK}.
}
\examples{
\dontrun{
 # assign to the line_search parameter in conj_grad:
 conj_grad(line_search = more_thuente(c1 = 0.05, c2 = 0.1), ...)
}
}
\references{
More, J. J., & Thuente, D. J. (1994).
Line search algorithms with guaranteed sufficient decrease.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{20}(3),
286-307.
}
\seealso{
This code is based on a translation of the original MINPACK code
 for Matlab by
 \href{https://www.cs.umd.edu/users/oleary/software/}{Dianne O'Leary}.
}

