% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasmussen.R
\name{rasmussen}
\alias{rasmussen}
\title{Rasmussen Line Search}
\usage{
rasmussen(c1 = c2/2, c2 = 0.1, int = 0.1, ext = 3, max_fn = 20)
}
\arguments{
\item{c1}{Constant used in sufficient decrease condition. Should take a value
between 0 and 1.}

\item{c2}{Constant used in curvature condition. Should take a value between
c1 and 1.}

\item{int}{Interpolation constant. Prevents step size being too small.}

\item{ext}{Extrapolation constant. Prevents step size extrapolation being
too large.}

\item{max_fn}{Maximum number of function evaluations allowed.}
}
\value{
Line search function for use in the \code{\link{conj_grad}} routine.
}
\description{
Line Search Factory Function
}
\details{
Line search algorithm originally written by Carl Edward Rasmussen in his
conjugate gradient routine. It consists of two main parts:
\enumerate{
 \item Using cubic extrapolation from an initial starting guess for the step
   size until either the sufficient decrease condition is not met or the
   strong curvature condition is met.
 \item Interpolation (quadratic or cubic) between that point and the start
   point of the search until either a step size is found which meets the
   Strong Wolfe conditions or the maximum number of allowed function
   evaluations is reached.
}

The extrapolation and interpolation steps are bounded at each stage to ensure
they don't represent too large or small a change to the step size.
}
\examples{
\dontrun{
# assign to the line_search parameter in conj_grad:
 conj_grad(line_search = rasmussen(c1 = 0.05, c2 = 0.1), ...)
}
}
\seealso{
Line search based on Matlab code by
 \href{http://learning.eng.cam.ac.uk/carl/code/minimize/}{Carl Edward Rasmussen}
 and also part of the Matlab
 \href{(http://www.gaussianprocess.org/gpml/code/matlab/doc/)}{GPML} package.
}

