% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polation.R
\name{tweak_extrapolation}
\alias{tweak_extrapolation}
\title{Tweak Extrapolated Point}
\usage{
tweak_extrapolation(xnew, x1, x2, EXT, INT)
}
\arguments{
\item{xnew}{1D position of the new point.}

\item{x1}{1D position of the first points used in the extrapolation.}

\item{x2}{1D position of the second point used in the extrapolation.}

\item{EXT}{Maximum multiple of \code{x2} that \code{xnew} is allowed to be
extrapolated to.}

\item{INT}{Given the distance between \code{x1} and \code{x2}, specified what
multiple of that distance is the minimum allowed distance for \code{xnew}
from \code{x2}.}
}
\value{
A value of \code{xnew} that obeys the minimum and maximum distance
 constraints from \code{x2}.
}
\description{
Prevents the extrapolated point from being too far away from or to close to
the points used in the extrapolation.
}

