% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polation.R
\name{tweak_interpolation}
\alias{tweak_interpolation}
\title{Tweak Interpolated Point}
\usage{
tweak_interpolation(xnew, x1, x2, INT)
}
\arguments{
\item{xnew}{Position of the interpolated point.}

\item{x1}{Position of the first point used for interpolation.}

\item{x2}{Position of the second point used for interpolation.}

\item{INT}{Given the distance between \code{x1} and \code{x2}, specifies what
multiple of that distance is the minimum allowed distance for \code{xnew}
from \code{x1} or \code{x2}.}
}
\value{
Tweaked position of \code{xnew} such that it is not too close to
either \code{x1} or \code{x2}.
}
\description{
Prevents interpolated point from getting too close to either of the
points used for the interpolation. If the point is not a number or infinite,
then it is set to the bisection of the position of the two interpolating
points before the check for a too-close approach is carried out.
}

